requirements.txt
playwright
requests
python-telegram-bot
